Creación de sidebar del panel de administración

+++ routes/web.php
@@ -26,13 +26,13 @@ Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name
  */
 
 // Logros
-Route::get('/admin/logros', [App\Http\Controllers\Administrador\LogrosController::class, 'index'])->name('logros.index');
-Route::post('/admin/logros', [App\Http\Controllers\Administrador\LogrosController::class, 'store'])->name('logros.store');
-Route::patch('/admin/logros/{id}/update', [App\Http\Controllers\Administrador\LogrosController::class, 'update'])->name('logros.update');
-Route::delete('/admin/logros/{id}/delete', [App\Http\Controllers\Administrador\LogrosController::class, 'destroy'])->name('logros.destroy');
+Route::get('/admin/logros', [App\Http\Controllers\Administrador\LogrosController::class, 'index'])->name('admin.logros.index');
+Route::post('/admin/logros', [App\Http\Controllers\Administrador\LogrosController::class, 'store'])->name('admin.logros.store');
+Route::patch('/admin/logros/{id}/update', [App\Http\Controllers\Administrador\LogrosController::class, 'update'])->name('admin.logros.update');
+Route::delete('/admin/logros/{id}/delete', [App\Http\Controllers\Administrador\LogrosController::class, 'destroy'])->name('admin.logros.destroy');
 
 // Géneros
-Route::get('/admin/generos', [App\Http\Controllers\Administrador\GenerosController::class, 'index'])->name('generos.index');
-Route::post('/admin/generos', [App\Http\Controllers\Administrador\GenerosController::class, 'store'])->name('generos.store');
-Route::PATCH('/admin/generos/{id}/update', [App\Http\Controllers\Administrador\GenerosController::class, 'update'])->name('generos.update');
-Route::delete('/admin/generos/{id}/delete', [App\Http\Controllers\Administrador\GenerosController::class, 'destroy'])->name('generos.destroy');
+Route::get('/admin/generos', [App\Http\Controllers\Administrador\GenerosController::class, 'index'])->name('admin.generos.index');
+Route::post('/admin/generos', [App\Http\Controllers\Administrador\GenerosController::class, 'store'])->name('admin.generos.store');
+Route::PATCH('/admin/generos/{id}/update', [App\Http\Controllers\Administrador\GenerosController::class, 'update'])->name('admin.generos.update');
+Route::delete('/admin/generos/{id}/delete', [App\Http\Controllers\Administrador\GenerosController::class, 'destroy'])->name('admin.generos.destroy');
